whizper_bot/requirements.txt
python-telegram-bot==21.6
tweepy>=4.14.0
requests>=2.32.0
aiohttp>=3.9.5
python-dotenv>=1.0.1
nest_asyncio>=1.6.0
fastapi>=0.111.0
uvicorn>=0.30.0
pandas>=2.2.2
numpy>=1.26.4
pytrends>=4.9.2
pyyaml>=6.0.2
